+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagUploadProguardTask.kt
@@ -110,5 +110,4 @@ open class BugsnagUploadProguardTask @Inject constructor(
             return project.tasks.register(name, configurationAction)
         }
     }
-
 }