+++ src/main/kotlin/com/bugsnag/android/gradle/internal/GradleUtil.kt
@@ -44,6 +44,8 @@ internal object AgpVersions {
     val VERSION_3_4: VersionNumber = VersionNumber.parse("3.4.0")
     val VERSION_3_5: VersionNumber = VersionNumber.parse("3.5.0")
     val VERSION_4_0: VersionNumber = VersionNumber.parse("4.0.0")
+    val VERSION_4_1: VersionNumber = VersionNumber.parse("4.1.0")
+    val VERSION_4_2: VersionNumber = VersionNumber.parse("4.2.0")
 }
 
 /** A fast file hash that don't load the entire file contents into memory at once. */
@@ -57,7 +59,7 @@ internal fun File.md5HashCode(): Int {
 }
 
 @Suppress("SpreadOperator")
-internal fun <T: Task> TaskProvider<out T>.dependsOn(vararg tasks: TaskProvider<out Task>): TaskProvider<out T> {
+internal fun <T : Task> TaskProvider<out T>.dependsOn(vararg tasks: TaskProvider<out Task>): TaskProvider<out T> {
     if (tasks.isEmpty().not()) {
         configure { it.dependsOn(*tasks) }
     }